requirements.txt
datetime
pandas
sqlalchemy
psycopg2-binary
python-decouple
